Issues have been solved
